c update
添加c#示例，更新c++代码，统一命名规范
