+++ .gitignore
@@ -1,2 +1,4 @@
 *.exe
 *.pdb
+.hintrc
+docs/*
