+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "*.c": "c",
+        "iosfwd": "cpp"
+    }
+}
