+++ README.md
@@ -1,16 +1,18 @@
-# 3xplus1
-用不同的编程语言验证3x+1问题 ,详情请看[这里](/intro.ipynb)
+# $3x+1$
+
+用不同的编程语言验证$3x+1$问题 ,详情请看[这里](/intro.ipynb)
 欢迎大家提交用不同的语言实现的代码~
 
-* [C](/threexp1.c)
-* [C++](/threexp1.cpp)
+* [C](/3xp1.c)
+* [C++ (mordern)](/3xp1.cpp)
+* [C#](/3xp1.cs)
 * [Crystal](/3xp1.cr)
 * [Go](/3xp1.go)
-* [JavaScript](/threexp1.js)
-* [Lua](/threexp1.lua)
-* [Nim](/threexp1.nim)
-* [Python](/threexp1.py)
-* [Ruby(和Crystal一样)](/3xp1.rb)
-* [Rust](/threexp1.rs)
+* [JavaScript](/3xp1.js)
+* [Lua](/3xp1.lua)
+* [Nim](/3xp1.nim)
+* [Python](/3xp1.py)
+* [Ruby (和 Crystal 一样)](/3xp1.rb)
+* [Rust](/3xp1.rs)
 * [文言](/%E4%B8%89%E5%80%8D%E5%85%83%E5%8A%A0%E4%B8%80.wy)
 * [逻辑电路仿真](/3xp1.CircuitProject)
